Remove UserContext
