+++ frontend/main-app/src/app/App.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { Route, Routes } from "react-router-dom";
 import "@digdir/design-system-tokens/brand/digdir/tokens.css";
 import classes from "./App.module.css";
@@ -14,9 +14,6 @@ import en from "../../../language/src/en.json";
 import { FormViewer } from "../components/FormViewer";
 import { SubHeader } from "../components/SubHeader";
 import { NotFound } from "../components/NotFound";
-// import { CreateNewForm } from "./CreateNewForm";
-import { User } from "../types";
-import { UserContext } from "../context";
 
 i18next.use(initReactI18next).init({
   resources: {
@@ -30,30 +27,26 @@ i18next.use(initReactI18next).init({
 });
 
 export const App = (): React.JSX.Element => {
-  const [userFromContext, setUserFromContext] = useState<User | null>(null);
-
   return (
-    <UserContext.Provider value={{ userFromContext, setUserFromContext }}>
-      <div className={classes.layout}>
-        <header>
-          <Header />
-          <SubHeader />
-        </header>
-        <main className={classes.main}>
-          <Routes>
-            <Route path="/" element={<Login />} />
-            <Route path="/signup" element={<Signup />} />
-            <Route path="/login" element={<Login />} />
-            <Route path="/login/:authError" element={<Login />} />
-            <Route path="/dashboard" element={<Dashboard />} />
-            <Route path="/form-builder" element={<FormBuilder />} />
-            <Route path="/form-builder/:formId" element={<FormBuilder />} />
-            <Route path="/view/:formId" element={<FormViewer />} />
-            <Route path="*" element={<NotFound />} />
-          </Routes>
-        </main>
-        <Footer />
-      </div>
-    </UserContext.Provider>
+    <div className={classes.layout}>
+      <header>
+        <Header />
+        <SubHeader />
+      </header>
+      <main className={classes.main}>
+        <Routes>
+          <Route path="/" element={<Login />} />
+          <Route path="/signup" element={<Signup />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="/login/:authError" element={<Login />} />
+          <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/form-builder" element={<FormBuilder />} />
+          <Route path="/form-builder/:formId" element={<FormBuilder />} />
+          <Route path="/view/:formId" element={<FormViewer />} />
+          <Route path="*" element={<NotFound />} />
+        </Routes>
+      </main>
+      <Footer />
+    </div>
   );
 };
