+++ frontend/main-app/src/components/Header/Header.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import { Header } from "./Header";
 import { MemoryRouter } from "react-router";
-import { useUser } from "../../hooks/useUser";
+import { useUserSession } from "../../hooks/useUserSession";
 
 const mockedUser = {
   id: "string",
@@ -12,8 +12,8 @@ const mockedUser = {
   organization: "string",
 };
 
-jest.mock("../../hooks/useUser");
-jest.mocked(useUser).mockReturnValue({ user: mockedUser, isLoading: false });
+jest.mock("../../hooks/useUserSession");
+jest.mocked(useUserSession).mockReturnValue({ user: mockedUser, isLoading: false });
 
 describe("Header component", () => {
   it("loads and displays Header Component", async () => {
