+++ frontend/main-app/src/components/Login/Login.tsx
@@ -1,7 +1,7 @@
 import classes from "./Login.module.css";
 import "@digdir/design-system-tokens/brand/digdir/tokens.css";
 import { Button, Heading, Textfield } from "@digdir/design-system-react";
-import React, { FormEvent, useContext, useState } from "react";
+import React, { FormEvent, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { validateLoginForm } from "./LoginUtils";
 import { getApiUrl } from "../../utils/getApiUrl";
@@ -9,13 +9,11 @@ import { LoginForm, LoginFormError } from "./types";
 import { useTranslation } from "react-i18next";
 import { alertToRender } from "../FormViewer/validationUtils";
 import axios from "axios";
-import { UserContext } from "../../context";
 
 export const Login = (): React.JSX.Element => {
   const { authError } = useParams();
   const navigate = useNavigate();
   const { t } = useTranslation();
-  const { setUserFromContext } = useContext(UserContext);
 
   const [fieldErrors, setFieldErrors] = useState<LoginFormError | null>(null);
   const [errorAlert, setErrorAlert] = useState(authError);
@@ -35,10 +33,9 @@ export const Login = (): React.JSX.Element => {
     const targetUrl = `${apiUrl}/users/login`;
 
     try {
-      const result = await axios.post(targetUrl, loginForm, {
+      await axios.post(targetUrl, loginForm, {
         withCredentials: true,
       });
-      setUserFromContext(result.data);
       navigate("/dashboard");
     } catch (error) {
       console.log(error);
