+++ frontend/main-app/src/components/Signup/Signup.tsx
@@ -1,7 +1,7 @@
 import classes from "./Signup.module.css";
 import "@digdir/design-system-tokens/brand/digdir/tokens.css";
 import { Button, Heading, Textfield } from "@digdir/design-system-react";
-import React, { FormEvent, useContext, useState } from "react";
+import React, { FormEvent, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { getApiUrl } from "../../utils/getApiUrl";
 import { validateSignupForm } from "./SignupUtils";
@@ -9,12 +9,10 @@ import { SignupForm, SignupFormError } from "./types";
 import { useTranslation } from "react-i18next";
 import { alertToRender } from "../FormViewer/validationUtils";
 import axios from "axios";
-import { UserContext } from "../../context";
 
 export const Signup = (): React.JSX.Element => {
   const navigate = useNavigate();
   const { t } = useTranslation();
-  const { setUserFromContext } = useContext(UserContext);
 
   const [fieldErrors, setFieldErrors] = useState<SignupFormError | null>(null);
   const [errorAlert, setErrorAlert] = useState("");
@@ -34,10 +32,9 @@ export const Signup = (): React.JSX.Element => {
     const targetUrl = `${apiUrl}/users`;
 
     try {
-      const result = await axios.post(targetUrl, signupForm, {
+      await axios.post(targetUrl, signupForm, {
         withCredentials: true,
       });
-      setUserFromContext(result.data);
       navigate("/dashboard");
     } catch (error) {
       console.log(error);
