+++ frontend/main-app/src/components/SubHeader/SubHeader.tsx
@@ -1,29 +1,25 @@
 import React, { useEffect, useState } from "react";
 import classes from "./SubHeader.module.css";
-import { Heading, Link } from "@digdir/design-system-react";
-import { ClipboardCheckmarkFillIcon, FilePlusFillIcon } from "@navikt/aksel-icons";
+import { Heading } from "@digdir/design-system-react";
 import { t } from "i18next";
-import { useLocation, useNavigate } from "react-router-dom";
-import { useUser } from "../../hooks/useUser";
-import { createNewForm } from "./utils";
+import { useLocation } from "react-router-dom";
+import { DashboardLink } from "./DashboardLink";
+import { formViewerHeading } from "./utils";
 
 export const SubHeader: React.FC = () => {
   const [heading, setHeading] = useState<string | React.JSX.Element>("");
-  const [links, setLinks] = useState<React.JSX.Element>(null);
   const pathname = useLocation().pathname;
 
   useEffect(() => {
     switch (true) {
       case pathname.startsWith("/login") || pathname === "/":
         setHeading(t("login_page.title"));
-        setLinks(null);
         break;
       case pathname.startsWith("/signup"):
         setHeading(t("signup_page.title"));
         break;
       case pathname.startsWith("/dashboard"):
         setHeading(t("dashboard"));
-        // setLinks(dashboardLinks());
         break;
       case pathname.startsWith("/view"):
         setHeading(formViewerHeading());
@@ -34,7 +30,7 @@ export const SubHeader: React.FC = () => {
     }
   }, [pathname]);
 
-  // if (pathname.startsWith("/form-builder/")) return;
+  if (pathname.startsWith("/form-builder/")) return;
 
   return (
     <div className={classes.subHeader}>
@@ -42,30 +38,8 @@ export const SubHeader: React.FC = () => {
         {heading}
       </Heading>
       <div className={classes.subHeaderLinks}>
-        {links}
-        {pathname.startsWith("/dashboard") && <DashboardLinks />}
+        {pathname.startsWith("/dashboard") && <DashboardLink />}
       </div>
     </div>
-  );
-};
-
-const DashboardLinks: React.FC = () => {
-  const navigate = useNavigate();
-  const { user } = useUser();
-
-  return (
-    <Link onClick={() => createNewForm(user, navigate)}>
-      {t("dashboard.new.form")}
-      <FilePlusFillIcon className={classes.subHeaderIcon} />
-    </Link>
-  );
-};
-
-const formViewerHeading = () => {
-  return (
-    <Link href="/">
-      {t("form_factory")}
-      <ClipboardCheckmarkFillIcon className={classes.subHeaderIcon} />
-    </Link>
   );
 };
