+++ frontend/main-app/src/context/context.ts
@@ -1,4 +1,3 @@
 import React, { createContext } from "react";
 
-export const UserContext = createContext(null);
 export const FormContext: React.Context<null | Form> = createContext(null);
