+++ frontend/main-app/src/hooks/useUserSession.tsx
@@ -3,9 +3,7 @@ import axios from "axios";
 import { User } from "../types";
 import { useEffect, useState } from "react";
 
-// Gets the user data by sending session data to backend
-
-export const useUser = () => {
+export const useUserSession = () => {
   const [user, setUser] = useState<User | null>(null);
   const [isLoading, setIsLoading] = useState(true);
